EY-1690: Trekkjer all den teknisk-generiske oppsettskoden ut i hjelpe-superklasse. Mistenkjer at denne på sikt kan trekkjast enda lengre ut, det meste av denne koden treng vi jo i alle integrasjonstestane våre
