+++ apps/etterlatte-behandling/src/test/kotlin/IntegrationTest.kt
@@ -1,12 +1,8 @@
 package no.nav.etterlatte
 
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
-import io.ktor.client.HttpClient
 import io.ktor.client.call.body
-import io.ktor.client.engine.mock.MockEngine
-import io.ktor.client.engine.mock.respond
 import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
-import io.ktor.client.request.HttpRequestBuilder
 import io.ktor.client.request.get
 import io.ktor.client.request.header
 import io.ktor.client.request.post
@@ -14,9 +10,6 @@ import io.ktor.client.request.setBody
 import io.ktor.http.ContentType
 import io.ktor.http.HttpHeaders
 import io.ktor.http.HttpStatusCode
-import io.ktor.http.fullPath
-import io.ktor.http.headersOf
-import io.ktor.serialization.jackson.JacksonConverter
 import io.ktor.serialization.jackson.jackson
 import io.ktor.server.testing.testApplication
 import no.nav.etterlatte.behandling.BehandlingDao
@@ -27,13 +20,7 @@ import no.nav.etterlatte.behandling.ManueltOpphoerResponse
 import no.nav.etterlatte.behandling.TilVilkaarsvurderingJson
 import no.nav.etterlatte.behandling.VedtakHendelse
 import no.nav.etterlatte.behandling.hendelse.HendelseDao
-import no.nav.etterlatte.behandling.klienter.GrunnlagKlient
-import no.nav.etterlatte.behandling.klienter.GrunnlagKlientTest
-import no.nav.etterlatte.behandling.klienter.VedtakKlient
-import no.nav.etterlatte.behandling.klienter.VedtakKlientTest
 import no.nav.etterlatte.behandling.objectMapper
-import no.nav.etterlatte.kafka.KafkaProdusent
-import no.nav.etterlatte.kafka.TestProdusent
 import no.nav.etterlatte.libs.common.behandling.BehandlingListe
 import no.nav.etterlatte.libs.common.behandling.BehandlingStatus
 import no.nav.etterlatte.libs.common.behandling.BehandlingType
@@ -42,7 +29,6 @@ import no.nav.etterlatte.libs.common.behandling.ManueltOpphoerAarsak
 import no.nav.etterlatte.libs.common.behandling.ManueltOpphoerRequest
 import no.nav.etterlatte.libs.common.behandling.Persongalleri
 import no.nav.etterlatte.libs.common.behandling.SakType
-import no.nav.etterlatte.libs.common.grunnlag.Grunnlag
 import no.nav.etterlatte.libs.common.grunnlag.Grunnlagsopplysning
 import no.nav.etterlatte.libs.common.gyldigSoeknad.GyldighetsResultat
 import no.nav.etterlatte.libs.common.gyldigSoeknad.GyldighetsTyper
@@ -54,56 +40,30 @@ import no.nav.etterlatte.libs.common.pdlhendelse.ForelderBarnRelasjonHendelse
 import no.nav.etterlatte.libs.common.pdlhendelse.UtflyttingsHendelse
 import no.nav.etterlatte.libs.common.soeknad.dataklasser.common.JaNeiVetIkke
 import no.nav.etterlatte.libs.common.tidspunkt.Tidspunkt
-import no.nav.etterlatte.libs.common.toJson
 import no.nav.etterlatte.libs.common.vilkaarsvurdering.VilkaarsvurderingUtfall
-import no.nav.etterlatte.libs.database.DataSourceBuilder
-import no.nav.etterlatte.libs.database.migrate
-import no.nav.etterlatte.libs.jobs.LeaderElection
 import no.nav.etterlatte.oppgave.OppgaveListeDto
 import no.nav.etterlatte.sak.Sak
-import no.nav.security.mock.oauth2.MockOAuth2Server
 import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertNotNull
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
-import org.testcontainers.containers.PostgreSQLContainer
-import org.testcontainers.junit.jupiter.Container
 import java.lang.Thread.sleep
 import java.time.Instant
 import java.time.LocalDate
 import java.time.LocalDateTime
 import java.time.YearMonth
 import java.util.*
-import javax.sql.DataSource
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-class ApplicationTest {
-
-    @Container
-    private val postgreSQLContainer = PostgreSQLContainer<Nothing>("postgres:12")
-    private val server: MockOAuth2Server = MockOAuth2Server()
-    private lateinit var beanFactory: TestBeanFactory
+class IntegrationTest : BehandlingIntegrationTest() {
 
     @BeforeAll
-    fun startServer() {
-        server.start(1234)
+    fun start() = startServer()
 
-        postgreSQLContainer.start()
-        postgreSQLContainer.withUrlParam("user", postgreSQLContainer.username)
-        postgreSQLContainer.withUrlParam("password", postgreSQLContainer.password)
-
-        beanFactory = TestBeanFactory(
-            jdbcUrl = postgreSQLContainer.jdbcUrl,
-            username = postgreSQLContainer.username,
-            password = postgreSQLContainer.password,
-            azureAdAttestantClaim = azureAdAttestantClaim,
-            azureAdSaksbehandlerClaim = azureAdSaksbehandlerClaim
-        ).apply { dataSource().migrate() }
-
-        beanFactory.behandlingHendelser().start()
-    }
+    @AfterAll
+    fun shutdown() = afterAll()
 
     @Test // TODO denne testen bør stykkes opp
     fun verdikjedetest() {
@@ -452,148 +412,4 @@ class ApplicationTest {
             HendelseDao { it }.finnHendelserIBehandling(behandlingOpprettet!!).also { println(it) }
         }
     }
-
-    @AfterAll
-    fun afterAll() {
-        server.shutdown()
-        postgreSQLContainer.stop()
-    }
-
-    private fun HttpRequestBuilder.addAuthToken(token: String) {
-        header(HttpHeaders.Authorization, "Bearer $token")
-    }
-
-    private val tokenSaksbehandler: String by lazy {
-        issueToken(
-            mapOf(
-                "navn" to "John Doe",
-                "NAVident" to "Saksbehandler01"
-            )
-        )
-    }
-
-    private val azureAdAttestantClaim: String by lazy {
-        "0af3955f-df85-4eb0-b5b2-45bf2c8aeb9e"
-    }
-
-    private val azureAdSaksbehandlerClaim: String by lazy {
-        "63f46f74-84a8-4d1c-87a8-78532ab3ae60"
-    }
-
-    private val tokenAttestant: String by lazy {
-        issueToken(
-            mapOf(
-                "navn" to "John Doe",
-                "NAVident" to "Saksbehandler01",
-                "groups" to listOf(
-                    azureAdAttestantClaim,
-                    azureAdSaksbehandlerClaim
-                )
-            )
-        )
-    }
-
-    private val tokenServiceUser: String by lazy {
-        issueToken(
-            mapOf(
-                "NAVident" to "Saksbehandler01",
-                "roles" to listOf("kan-sette-kilde") // TODO hva brukes dette til?
-            )
-        )
-    }
-
-    private fun issueToken(claims: Map<String, Any>) =
-        server.issueToken(
-            issuerId = ISSUER_ID,
-            audience = CLIENT_ID,
-            claims = claims
-        ).serialize()
-
-    private companion object {
-        const val ISSUER_ID = "azure"
-        const val CLIENT_ID = "mock-client-id"
-    }
-}
-
-class TestBeanFactory(
-    private val jdbcUrl: String,
-    private val username: String,
-    private val password: String,
-    private val azureAdSaksbehandlerClaim: String,
-    private val azureAdAttestantClaim: String
-) : CommonFactory() {
-    override fun getSaksbehandlerGroupIdsByKey(): Map<String, String> =
-        mapOf(
-            "AZUREAD_ATTESTANT_GROUPID" to azureAdAttestantClaim,
-            "AZUREAD_SAKSBEHANDLER_GROUPID" to azureAdSaksbehandlerClaim
-        )
-
-    val rapidSingleton: TestProdusent<String, String> by lazy { TestProdusent() }
-    override fun dataSource(): DataSource =
-        DataSourceBuilder.createDataSource(jdbcUrl, username, password)
-
-    override fun rapid(): KafkaProdusent<String, String> = rapidSingleton
-
-    override fun vedtakKlient(): VedtakKlient {
-        return VedtakKlientTest()
-    }
-
-    override fun grunnlagKlient(): GrunnlagKlient {
-        return GrunnlagKlientTest()
-    }
-
-    override fun pdlHttpClient(): HttpClient = HttpClient(MockEngine) {
-        engine {
-            addHandler { request ->
-                if (request.url.fullPath.startsWith("/")) {
-                    val headers = headersOf("Content-Type" to listOf(ContentType.Application.Json.toString()))
-                    val json = javaClass.getResource("")!!.readText() // TODO: endre name
-                    respond(json, headers = headers)
-                } else {
-                    error(request.url.fullPath)
-                }
-            }
-        }
-        install(ContentNegotiation) {
-            register(
-                ContentType.Application.Json,
-                JacksonConverter(no.nav.etterlatte.libs.common.objectMapper)
-            )
-        }
-    }
-
-    override fun grunnlagHttpClient(): HttpClient = HttpClient(MockEngine) {
-        engine {
-            addHandler { request ->
-                if (request.url.fullPath.startsWith("/")) {
-                    val headers = headersOf("Content-Type" to listOf(ContentType.Application.Json.toString()))
-                    respond(Grunnlag.empty().toJson(), headers = headers)
-                } else {
-                    error(request.url.fullPath)
-                }
-            }
-        }
-        install(ContentNegotiation) {
-            register(
-                ContentType.Application.Json,
-                JacksonConverter(no.nav.etterlatte.libs.common.objectMapper)
-            )
-        }
-    }
-
-    override fun leaderElection() = LeaderElection(
-        electorPath = "electorPath",
-        httpClient = HttpClient(MockEngine) {
-            engine {
-                addHandler { req ->
-                    if (req.url.fullPath == "electorPath") {
-                        respond("me")
-                    } else {
-                        error(req.url.fullPath)
-                    }
-                }
-            }
-        },
-        me = "me"
-    )
 }
